+++ widgets/file.py
@@ -291,7 +291,7 @@ class FileWidget(TreeWidget):
 
 	@staticmethod
 	def render(moduleName, adminInfo, context):
-		rootNode = context.get("rootNode") if context else None
+		rootNode = context.get(conf["vi.context.prefix"] + "rootNode") if context else None
 		return FileWidget(module=moduleName, rootNode=rootNode, context=context)